road-sense-service/requirements.txt
python-dotenv
celery
redis
Django
django-celery-results
pydantic-settings
pydantic
scikit-learn
numpy
joblib
django-cors-headers
channels 
channels_redis
daphne
djangorestframework
djangorestframework-simplejwt
python-dotenv




# Django==4.2.7
# djangorestframework==3.14.0
# django-cors-headers==4.3.1
# channels==4.0.0
# channels-redis==4.1.0
# celery==5.3.4
# redis==5.0.1
# psycopg2-binary==2.9.7
# dj-database-url==2.0.0
# gunicorn==21.2.0
# whitenoise==6.5.0
# scikit-learn==1.3.2
# pandas==2.1.3
# numpy==1.25.2
# joblib==1.3.2